inpaint-service/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
# PyTorch will be installed from official wheels in Dockerfile/Dockerfile.gpu
# torch==2.4.1
# torchvision==0.19.1
diffusers==0.25.0
transformers>=4.36.0
tokenizers>=0.15.0
accelerate==0.24.0
numpy==1.24.3
huggingface-hub==0.19.4
simple-lama-inpainting>=0.1.0